Split TrailingCommaRule
