+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -2,7 +2,6 @@ package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
-import com.pinterest.ktlint.ruleset.experimental.trailingcomma.TrailingCommaRule
 
 public class StandardRuleSetProvider : RuleSetProvider {
 
@@ -51,7 +50,8 @@ public class StandardRuleSetProvider : RuleSetProvider {
         SpacingBetweenDeclarationsWithAnnotationsRule(),
         SpacingBetweenDeclarationsWithCommentsRule(),
         StringTemplateRule(),
-        TrailingCommaRule(),
+        TrailingCommaOnDeclarationSiteRule(),
+        TrailingCommaOnCallSiteRule(),
         WrappingRule()
     )
 }
